Python package for badactors
Automatic deployment on PyPi with travis and install with pip.
