+++ README.md
@@ -1,22 +1,30 @@
 # badactors
-badactors.py will create a list of about 51,000 de-duplicated bad actors from public IP blacklist for your infosec needs. 
+badactors.py will create a list of about 51,000 de-duplicated bad actors from
+public IP blacklist for your infosec needs. 
 
-reallybadactors.py will create a list of about 87,000 de-duplicated bad actors from public IP blacklist for your big data infosec needs.
+reallybadactors.py will create a list of about 87,000 de-duplicated bad actors
+from public IP blacklist for your big data infosec needs.
 
 *(reallybadactors.py takes about 25 minutes to run. So...)*
 
 ![Alt text](https://imgs.xkcd.com/comics/compiling.png)
 
+## Install
+
+```shell
+$ pip install badactors
+```
+
 ## Usage
-python badactors.py
-or
-python reallybadactors.py
+
+```shell
+$ badactors
+$ # or
+$ reallybadactors
+```
 
 ## ScreenShot
 ![Alt text](http://40.media.tumblr.com/4a71ec0ce7c50b8bbac13744f63d96bb/tumblr_inline_nohy84LkjN1qf8s52_500.png)
 
 ## Important Notice
 I dont know what I am doing. It isnt perfect but it works. 
-
-
-
