+++ setup.cfg
@@ -1,0 +1,41 @@
+[metadata]
+name = badactors
+home-page = https://github.com/jgamblin/badactors
+summary = Create a list of bad actors from public IP blacklist.
+description-file = 
+    README.md
+author = Jerry Gamblin
+author-email = jerry.gamblin@gmail.com
+licence = MIT
+classifier =
+    Development Status :: 5 - Production/Stable
+    Intended Audience :: Information Technology
+    License :: OSI Approved :: MIT License
+    Operating System :: POSIX
+    Programming Language :: Python
+    Programming Language :: Python :: 2
+    Programming Language :: Python :: 2.7
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.4
+    Programming Language :: Python :: 3.5
+    Programming Language :: Python :: 3.6
+
+[files]
+packages =
+    badactors
+
+[extras]
+devel=
+    fixtures
+    pbr
+    tox
+    bandit
+
+[entry_points]
+console_scripts =
+    badactors = badactors.badactors:main
+    reallybadactors = badactors.reallybadactors:main
+    twostrikes = badactors.twostrikes:main
+
+[wheel]
+universal = 1
