+++ setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup
+
+setup(
+    requires=['pbd'],
+    pbr=True,
+    long_description_content_type="text/markdown",
+)
